+++ Scripts/MMOGame/Database/MySQL/MySQLDatabase_StorageItem.cs
@@ -33,12 +33,12 @@ namespace MultiplayerARPG.MMO
             {
                 result = new CharacterItem();
                 result.dataId = reader.GetInt32("dataId");
-                result.level = (short)reader.GetInt32("level");
-                result.amount = (short)reader.GetInt32("amount");
+                result.level = reader.GetInt16("level");
+                result.amount = reader.GetInt16("amount");
                 result.durability = reader.GetFloat("durability");
                 result.exp = reader.GetInt32("exp");
                 result.lockRemainsDuration = reader.GetFloat("lockRemainsDuration");
-                result.ammo = (short)reader.GetInt32("ammo");
+                result.ammo = reader.GetInt16("ammo");
                 result.sockets = ReadSockets(reader.GetString("sockets"));
                 return true;
             }
